Avoid 'conversion may change value' cases in listpanel.cpp and krlayoutfactory.cpp
Avoid three warnings in listpanel.cpp and krlayoutfactory.cpp, confirming the
conversions to `int` because three Qt functions that are used (`setValue(int
value)`, `qUncompress(const uchar *data, int nbytes)` and `QByteArray(const
char *data, int size)`) only accept `int` arguments.

Revision: https://invent.kde.org/utilities/krusader/-/merge_requests/56

+++ krusader/Panel/krlayoutfactory.cpp
@@ -93,9 +93,9 @@ bool KrLayoutFactory::parseResource(const QString& path, QDomDocument &doc)
     if (res.isValid()) {
         QByteArray data;
         if (QRESOURCE_ISCOMPRESSED(res))
-            data = qUncompress(res.data(), res.size());
+            data = qUncompress(res.data(), static_cast<int>(res.size()));
         else
-            data = QByteArray(reinterpret_cast<const char*>(res.data()), res.size());
+            data = QByteArray(reinterpret_cast<const char *>(res.data()), static_cast<int>(res.size()));
         return parseContent(data, path, doc);
     } else {
         qWarning() << "resource does not exist:" << path;
